test: it runs
