+++ .idea/Computing-Group-Project.iml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/ai_friend/.dart_tool" />
+      <excludeFolder url="file://$MODULE_DIR$/ai_friend/.pub" />
+      <excludeFolder url="file://$MODULE_DIR$/ai_friend/build" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
