+++ ai_friend/.dart_tool/package_config.json
@@ -170,7 +170,7 @@
       "languageVersion": "3.6"
     }
   ],
-  "generated": "2025-03-13T05:30:53.383155Z",
+  "generated": "2025-03-14T05:38:28.004099Z",
   "generator": "pub",
   "generatorVersion": "3.6.1",
   "flutterRoot": "file:///C:/src/flutter",
