+++ ai_friend/lib/pages/diary-inside.dart
